+++ assets/js/admin-builder-stripe-card-field.js
@@ -78,7 +78,7 @@ var WPFormsStripeCardField = window.WPFormsStripeCardField || ( function( docume
 		 */
 		paymentsEnabledCheck: function() {
 
-			if ( ! $( '.wpforms-field.wpforms-field-' + wpforms_builder_stripe_card_field.field_slug ).length ||
+			if ( ! $( `.wpforms-field.wpforms-field-${ wpforms_builder_stripe_card_field.field_slug }:visible` ).length ||
 				$( '#wpforms-panel-field-stripe-enable' ).is( ':checked' ) ) {
 				return;
 			}